Rewrite AI tab for Codex-style speed and minimal UX.
Split the 2.6k-line Insights view into focused modules so typing and new-chat stay off the hot path, while keeping local history search and Cmd/Ctrl+K palette behavior unchanged.
